feat: handle setting up outlines for multiple shaders - adds support for Nya222 HSR geometry nodes setup
